feature: delete link from collection
